chore: fmt

+++ src/adapters/MorphoMarketV1Adapter.sol
@@ -96,7 +96,7 @@ contract MorphoMarketV1Adapter is IMorphoMarketV1AdapterStaticTyping {
         require(burnSharesExecutableAt[marketId] != 0, NotTimelocked());
         require(block.timestamp >= burnSharesExecutableAt[marketId], TimelockNotExpired());
         burnSharesExecutableAt[marketId] = 0;
-        uint supplySharesBefore = positions[marketId].supplyShares;
+        uint256 supplySharesBefore = positions[marketId].supplyShares;
         positions[marketId].supplyShares = 0;
         emit BurnShares(marketId, supplySharesBefore);
     }
